refactor(popover): colocate tag item component
