+++ frontend/src/components/popover/edit-filetag-popover/tag-item.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { seafileAPI } from '../../utils/seafile-api';
-import { Utils } from '../../utils/utils';
-import Icon from '../icon';
-import toaster from '../toast';
+import { seafileAPI } from '../../../utils/seafile-api';
+import { Utils } from '../../../utils/utils';
+import Icon from '../../icon';
+import toaster from '../../toast';
 
 class TagItem extends React.Component {
